ap.rpcd shouldn't report incorrect release
Summary:
If you upgrade just the packages, the link to the release descriptor in
`/opt/com.brightgate/etc` is no longer correct, but still exists.  Get
rid of it in the package postinstall script, for that case; the
installer code will re-create it correctly.

This also collects some information about the versions of the artifacts
installed on the currently-running system, and reports that to the
cloud, which (for now) simply logs it.

Fixes T456

Test Plan:
Run on BG-100, development cloud, upgrade packages independently, and see the
eventd log confirm that it's not recording the release (unknown UUID), but
noting the commits it was able to discover on the appliance.

Reviewers: dp, stephen

Reviewed By: dp

Maniphest Tasks: T456

Differential Revision: https://ph0.b10e.net/D794