Silence the info header when re-running make
Summary:
With the automation of the Go module makefile dependency analysis,
`make` restarts after that computation, which is pretty much every time.
This is noisy and annoying because of the large informational header we
print on every invocation.  The `make` variable `$(MAKE_RESTARTS)` can
be checked to see how many times we've restarted, and only print that
header the first time.

Test Plan:
Run make in various scenarios and note that the header only prints out once, at
the very beginning.

Reviewers: dp, stephen

Reviewed By: dp

Differential Revision: https://ph0.b10e.net/D795