Write fault records to GCS
Summary:
We should be archiving faults to cloud storage as we are other events.
This enables that, but doesn't remove the code that writes them to the
filesystem.

Test Plan:
Crash a daemon to generate a fault.  See that on the cloud side the fault is
written both to disk and to the correct bucket.

Reviewers: dp, nils

Reviewed By: dp, nils

Differential Revision: https://ph0.b10e.net/D796