Daemons can trigger MemoryFaults, including heap profile
Summary:
ap.mcp will set up environment variables with a daemon's memory limits
before launching the daemon.  The daemon will monitor its memory
consumption over time.  When it exceeds its Warn threshhold, it will
increase its memory profile rate.  When it is halfway between the Warn
and Kill threshholds, it will generate a heap profile, and will trigger
a MemoryFault report which includes that profile in the payload.

Test Plan:
added a memory leak to logd, and verified that the expected
report was triggered.  Lowered the memory limits of watchd signficantly,
again verifying that the fault was recorded.

Differential Revision: https://ph0.b10e.net/D801