Account info fetch isn't working right for non-admin users
Summary:
This fixes a bug that was encountered by a non-admin account; because
the user is at a low level of privilege, they cannot get basic account
information, even about themselves.  This causes an assertion failure in
the account_prefs.vue page.

The fix is to allow those with "user" privilege the ability to fetch
basic account info for themselves; the results of
/api/org/:org_uuid/accounts is filtered to just return the user
in question.

Test Plan: Tested against "admin" and "user" roled accounts in simulated cloud

Reviewers: duvall, stephen

Reviewed By: duvall

Differential Revision: https://ph0.b10e.net/D803