T460 default_pass and vulnerabilities should be independently cleared
Summary:
Allow individual tests to report errors as well as clear/vulnerable.

After running a full vulnerability scan, we should not treat the absence of a
result for one test as if that vulnerability has been addressed.  We should only
clear a previously detected vunerability if we see an explicit indication that
that specific vulnerability test has been re-run successfully.

Test Plan:
manual on nils-wrt with a pi with both Exim and default password
vulnerabilities.

Differential Revision: https://ph0.b10e.net/D808