Add Alpha 2 content for accounts/roles and orgs/sites
Summary:
This review is less than ideal because it is a bundle of things:
1. Accounts/roles content (user management)
2. Org/site concepts (perhaps overly simplified, so feedback very welcome)
3. Addition of Support page information
4. change to the language we use for the expected network name (if admins don't adopt the -users suffix)

I've reviewed the screen shots in the current guide and the only one to retake is the Device Details shot, but there's not good data for that right now, so I've not yet retaken it.

I'd like to add a shot of the new switcher for orgs, but that's in code review, so will add this later.

Feedback, as always, very welcome.

Test Plan: Tested links and layout using HTML Preview package in Atom

Reviewers: dp, duvall, stephen, cat

Reviewed By: dp

Differential Revision: https://ph0.b10e.net/D810