ap-factory shouldn't error out without -d
Summary:
When `ap-factory` is run without `-d` (such as when following the field
manual), the path to the artifacts it's installing defaults to `.`, and
`filepath.Rel()` fails, causing `ap-factory` to exit with a usage
message and a non-zero exit code.  Make sure to create an absolute path
to use as the source directory to avoid this problem.  Also don't let
any failures during release symlink processing cause `ap-factory` to
exit unsuccessfully.

We should also not remove the symlink on this side when installing onto
the other side.

Test Plan:
Run `ap-factory` both without `-d` and with `-d ./some/path` and see that it
completes without error.

Reviewers: dp, stephen

Reviewed By: dp, stephen

Differential Revision: https://ph0.b10e.net/D813