Change the git commit hash collection to deal with tags
Summary:
Now that we've started tagging our releases, the mechanism used to retrieve the
long-form git commit hash used to locate artifacts in cloud storage no longer
gives the same results, giving just the tag name, not the hash.  The short form
is also different, but doesn't need to change.  The release reporting code will
eventually need to change to send just the hash (and -dirty), but nothing parses
that yet.

Test Plan: Run the new command manually to make sure it outputs what we expect.

Reviewers: dp

Reviewed By: dp

Differential Revision: https://ph0.b10e.net/D814