Add GetNodes() to cfgapi and "get nodes" to configctl
Summary:
GetNodes() provides a single, convenient call to collect information
from the @/nodes, @/network, and @/metrics subtrees, constructing a
high-level picture of all the node in a cluster.

"<cl|ap>-configctl get nodes" consumes that information and presents it
in a user-friendly format.

Test Plan:
cl-configctl tested on build0, ap-configctl tested on
nils-wrt

Differential Revision: https://ph0.b10e.net/D823