If an OpenWRT box has a missing/invalid serial no, generate one
Summary:
If we fail to find a valid serial number in the firmware or filesystem,
generate a new one with a special, reserved Site Code (ZZ).  If rpcd
discovers that its node has one of these serial numbers, send an event
to the cloud, to indicate that the appliance needs maintenance.

Test Plan: manual on nils-wrt

Differential Revision: https://ph0.b10e.net/D824