ChoosePort() should test for availability within range
Summary:
`network.ChoosePort()` chooses an ephemeral port number by binding to
port 0 and checking what port actually got used.  This is useful for
some of our uses of the function, but when choosing a port for the ssh
tunnel, it needs to be within the range of ports allowed by the
firewall.

This change modifies the function to allow it to take zero, one, or two
arguments, and choose an ephemeral port, the given port, or a port from
the range, respectively.  It will fail if no allowed port is available.

This also changes `cl-service` to choose a port from the allowed range,
so that we no longer have to guess a port and specify it by hand.

Fixes T321

Test Plan:
Run new `cl-service` on `tun0` without the `-ssh_port` flag and see that it
always picks an unused port.  With the flag and a port that's already in use,
see that it dies when it recognizes that fact.

Reviewers: nils

Reviewed By: nils

Maniphest Tasks: T321

Differential Revision: https://ph0.b10e.net/D829