Auto-detect external IP addresses in a GCP VM
Summary:
We can use the metadata server available to every GCP VM to discover our
external IP address for the incoming SSH connection to use.  This means
we shouldn't have to use the `-ssh_addr` flag to `cl-service` except in
extraordinary circumstances.

Test Plan:
Run the new `cl-service` on `tun0` without the `-ssh_addr` flag and see that
everything works.

Reviewers: dp, nils

Reviewed By: dp, nils

Differential Revision: https://ph0.b10e.net/D831