Restore active daemon set on mcpd restart
Summary:

ap.mcp and the remaining daemons were managed as independent services.
If ap.mcp crashes, it would be automatically restarted by its service,
but the brightgate service was unaware that all of the child daemons
need to be restarted as well.

ap.mcp will now track which daemons are intended to be running,
persisting that list to a file in /tmp.  When ap.mcp starts, it will
consult that file and will attempt to restart any daemons that were in
the active set when ap.mcp was last running.  If the file doesn't exist,
all daemons will be started.

Because the state file is stored in /tmp, it will be cleared on reboot.
When the mcp service is shut down cleanly, the state file will be
removed.  This is intended to ensure that on a reboot, all of the
daemons will be restarted.

Remove brightgate-appliance service.

Test plan: manual testing on nils-wrt.  Tested daemon crashes, clean
shutdowns, and reboots.

Differential Revision: https://ph0.b10e.net/D834