Update ap-speedtest to use latest module, latest spec
Summary:
The servers we use updated to the 0.8.0 version of the NDT7 spec, which
means our old clients no longer work properly.  This change updates the
NDT7 client library to a version that understands the new spec, and
reworks our code to comply.

Test Plan: Run `ap-speedtest` on x86 and the BG-100 to see that we get reasonable numbers.

Reviewers: dp

Reviewed By: dp

Differential Revision: https://ph0.b10e.net/D835