ap-factory check environment validity on install, suggest passwords
Summary:
This changeset causes `ap-factory` to check the values on the boot and
bootmenu entries in the U-Boot environment.  It also will suggest five
passwords on install (or when given the `passwd` subcommand), as part of
our "simplify factory install" work.

Test Plan:
Built and tested on MT7623 (although mostly with Beta-1 distro
candidate).

Reviewers: dp, duvall

Differential Revision: https://ph0.b10e.net/D836