Remove ap.serviced dependency on satellite daemons
Summary:
The daemons running on the satellite need to communicate with ap.configd
running on the gateway.  To find that node's IP address, we were doing a
DNS lookup for 'gateway'.  This lookup can only succeed if ap.serviced
is running on the gateway.  We don't want the satellite stack to rely on
that much of the gateway stack to be functional, so this change
eliminates that DNS lookup.

Before launching a daemon, ap.mcp will set its APGATEWAY environment
variable to the IP address of the gateway node.  This means satellite
daemons will not need to perform a DNS lookup.

ap.mcp has a new GATEWAY command, which will return the IP address of
the gateway node.  This allows CLI tools (eg, ap-configctl) to get the
gateway address from ap.mcp without being a child of it.  This change
removes the need for those tools to examine the DHCP lease info directly
to find the gateway node, which means that they can now be run by
non-root users.

The different communications timeouts were not being propagated to the
underlying mangos socket layer.  This was masking short-term
(sub-minute) mcp outages and quick restarts, which was preventing a
timely shutdown/restart of the satellite daemons.

After crashing N times, a daemon will not be automatically restarted for
N seconds.  If the failure was caused by some transient state, this
delay increases the chances that that transient state will be cleared,
finally allowing the daemon to restart successfully.  The number of
times we will attempt to restart a daemon is increased from 4 to 10,
which also increases the likelihood of a transient failure being
resolved.

Test Plan:
manual testing on two openwrt boards:
   - upgrading the bg-appliance package in different orders
   - booting satellite before the gateway and rebooting the nodes in
     different orders
   - pulling the network cable between satellite and gateway before and
     after the mcp peer relationship was established
   - restarting ap.mcp, configd, and network daemons on both nodes

Differential Revision: https://ph0.b10e.net/D851