Congestion-aware channel selection
Summary:
We monitor how many other APs are using each channel, and use that
information to decide which channel we should use to get the least
interference.  We will reexamine that choice periodically (by default
every 3 hours), and change if our current channel(s) now appear to be
congested.

Test Plan:
manual, using an artificially low "reevaluate" setting to
observe the AP switching channels to react to changes in the
environment.

Differential Revision: https://ph0.b10e.net/D860