T474 hostname should follow node ID
Summary:
At startup, if the /data/mcp/serial file doesn't exist, try to create it
with the serial number reported by fw_printenv.

If the system's serial number is defined, use it as the hostname.  If
there is no serial number, fall back to the legacy machineid.  If that
fails, use 'unknown'.

Test Plan:
manual testing across reboots, with the file both present and
missing.

Differential Revision: https://ph0.b10e.net/D861