ap.mcp should be less aggressive about finding a DHCP address
Summary:
Rather than examining every interface in the system for a DHCP address,
we should only look at those interfaces configured to ask for one.

The list of DHCP-enabled interfaces is likely to be static post-boot, so
it doesn't need to be rebuilt/rediscovered on each iteration.  Instead,
we only reevalulate the list of potential interfaces once a minute.

Rather than polling for a DHCP every second in perpetuity, gradually
increase the delay between polls up to a minute.

Test Plan:
Tunning on use on my test board in a DHCP-configured installation.  Running
on dogfood in a static IP installation.  Running in the conference room as
a DHCP/satellite installation.

Differential Revision: https://ph0.b10e.net/D863