Disallow setState() for pseudo devices
Summary:
A wireless NIC's pseudo partners (wlanX_*, used for alternate SSIDs)
share a wifiInfo struct with the physical device.  Config tree changes
to the pseudo devices should be ignored and the setState() routine
should only be invoked on the physical device.  This prevents ambiguity
if the config tree contains different states for the physical device and
its partners.  In practice, this should only happen when attempting to
disable/enable a device, as no other states should be set by the user.

Test Plan:
manual testing, repeatedly toggling multiple devices between
'disabled' and 'ok'.

Differential Revision: https://ph0.b10e.net/D869