Reset retransmit count on successful connection
Summary:
We track how many times a client triggers a CTRL-EVENT-EAP-RETRANSMIT
events.  One or two occurences on a connection attempt are not unusual.
If we see more than that, we assume the client is stuck in a dead-end
and kick it off the network to force a complete reauthentication.

If a client disconnects and reconnects quickly enough, it can exceed
this RETRANSMIT threshold - even if it connects successfully between
reconnection attempts.

This change resets the count of RETRANSMIT events when a client
successfuly connects.  This prevents us from taking drastic action to
correct a client that doesn't need correction.

Test Plan:
Testing initial connection attempts with a script that turns
a Mac client's wifi on and off every minute.  Testing reauthentication
attempts by setting hostap's "eap_reauth_period" to 60.

Differential Revision: https://ph0.b10e.net/D870