T499 Orphaned clients may remain marked active even though they aren't
Summary:
Whenever hostapd is shut down, all wireless clients connected to that
node will be disconnected.  This may happen for hostapd crashes or
restarts, or when ap.networkd crashes or restarts.  Each time we start
hostapd on a node, we iterate over the full client list and set the
'active' property to false for each wireless client that was last seen
connected to that node, or to a node that we no longer recognize.

Test Plan:
manual testing with live clients and stale clients.  Also
modified clients' connection/node properties to simulate lost nodes.

Differential Revision: https://ph0.b10e.net/D876