dhcp_name is descriptive - not prescriptive
Summary:
We will accept and retain for diagnostic purposes the hostname requested
by clients during DHCP, but we will not use that name to create DNS
entries.

Because this name is no longer being used for identification, we have
dropped the uniqueness requirement.  There may be any number of clients
with the same dhcp_name.  The existence of a client with a given
dhcp_name will not prevent any other client from having a dns_name of
the same value.

The client cfgapi call DisplayName() will still return the DHCPName if
DNSName is not set.

Test Plan: manual

Differential Revision: https://ph0.b10e.net/D879