Add support for "friendly" names
Summary:
This adds two new per-client properties: friendly_name and friendly_dns.

friendly_name is an arbitrary string used to identify a device.  It is
expected to be set by the user with the web interface, and will
subsequently be displayed in the interface to allow the user to identify
the device.

friendly_dns is a legal DNS name, derived by the appliance from the
friendly_name property.  friendly_dns will be unique with respect to
other derived DNS names and explicitly set dns_name properties.  We will
allow the dns_friendly property to be set using the CLI, but it is
discouraged.  If the user wants a DNS name other than the one we derive,
they should set the client's dns_name property.

Test Plan: manual testing with multple clients

Differential Revision: https://ph0.b10e.net/D882