cl-obs use cloud storage for source data; clean up training set to match
Summary:
This changeset adds the ability for `cl-obs` to use cloud storage as its
source for DeviceInfo files.  Moving to that source means that the
training set needed to be cleaned up to match, which required deleting
references to the oldest records.  Significant DNS queries for some
devices were added to the defined set of interest.  TCP and UDP ports
above 10000 are now ignored for training.  Redundant training records
have been removed.

A sample script for `cron`-based execution is included.

Test Plan: Built; run repeatedly.

Reviewers: dp, duvall, nils

Differential Revision: https://ph0.b10e.net/D858