Add support for load balancing to cl.httpd, cl.rpcd
Summary:
The cloud services need some tweaks to be run manageably behind a load balancer.

`cl.httpd` gets a big config switch, `B10E_CLHTTPD_LOAD_BALANCED`, to switch to
this mode, and a smaller one `B10E_CLHTTPD_ALLOWED_HOSTS`, which lets us run it
on arbitrary hosts.  We add a route for a basic health check.  We also add some
LB-related information to the log.

`cl.rpcd` needs to serve a health check outside of its gRPC services, and also
requires a certificate for the load-balancer proxies to talk to it, so we allow
for it to generate one on the fly.

We also add support for `cl.rpcd` and the appliance stack to dump their session
keys to a file for ease in debugging the sessions.

Test Plan:
Run `cl.httpd` and `cl.rpcd` in instance groups behind load balancers, ensuring
that the browser UI functionality remains intact, and that gRPC requests are
answered as expected.

Reviewers: dp

Reviewed By: dp

Differential Revision: https://ph0.b10e.net/D883