ap.mcp restart backoff not working as intended
Summary:
If a daemon crashes repeatedly, we want to insert an increasingly large
delay between restart attempts.  If the failure is caused by an external
condition, this increases the odds that the condition is resolved.  If
the failure is caused by instability in another daemon, this gives that
daemon additional time to stabilize before interacting with it.  If the
failure is caused by general system load, this spreads out the naturally
occuring load spikes when each daemon starts.

The current code calculates the "next restart time" from the time the
daemon was last started, when we really want it to be calculated from
the time the daemon shuts down.

Test Plan: manual on nils-wrt

Differential Revision: https://ph0.b10e.net/D893