T505 existing leases seem to override static IP assignments
Summary:
When setting a static IP address, we were adding the new entry to the
DHCP lease table, but we were not clearing the old one.  When the client
issued another DHCP request, it would be matched with whichever IP
address appeared first in the table.

Also:
    - Now detect and report multiple leases for a single mac
    - Stop issuing the PROVISION event that nobody listens for
    - Make the code flow through dhcpIPv4Changed() easier to follow

Test Plan: manual testing on nils-wrt

Differential Revision: https://ph0.b10e.net/D898