T386 appliance DNS server rejects requests originating from the appliance
Summary:
This adds 127.0.0.1 as a valid DNS requestor, with a nominal ring
assignment of CORE.  This allows ap.serviced to respond to DNS requests
originating from the appliance for local or remote hostnames.  This will
not support requests from the appliance to the ring router address
(e.g., 192.168.5.1) ; it will only support requests to the localhost
address.

Test Plan:
tested resolving both local and remote hostnames when
specifying the nameserver explicitly ("dig @127.0.0.1") and implicitly
(with a modified /etc/resolv.conf).

Differential Revision: https://ph0.b10e.net/D903