Switch local stats format from JSON to GOB
Summary:
Currently stats are persisted on the appliance in JSON.  After being
uploaded to the cloud, the daily merged files are persisted in the Go
Binary format.  The binary files are ~75% smaller than JSON and require
much less CPU to process.

With this change, local data will be persisted in the binary format.
cl.rpcd will allow the upload of either JSON or GOB archives.

Test Plan: currently running on nils-wrt and svc0.b10.net

Differential Revision: https://ph0.b10e.net/D907