Avoid negative timeouts in the comms layer
Summary:
An ap.mcp on a satellite was hanging for over a week in:
	satelliteLoop()->connectToGateway()-> ... -> mangos.sendMsg()

There's not enough information in the SIGABRT-triggered stack dump to be
completely certain, but I suspect the problem is that we set a negative
SendDeadline at comms.go/Send():247.  An inspection of the mangos code
shows that a negative deadline is handled the same as no deadline at
all, which would lead to an infinite hang.  The implemented fix ensures
that we always set a positive timeout before calling into the mangos
code.

Test Plan:
I haven't been able to reproduce the original problem, so I
can't be sure this will fix it.  I have verified that this change
doesn't break anything.

Differential Revision: https://ph0.b10e.net/D908