Replace prometheus on the appliance with @/metrics
Summary:
This adds a basic bgmetrics module, which replaces prometheus.  Rather
than spawning a web server in each daemon to feed the collected stats to
a non-existent prometheus daemon, we now periodically send updated stats
to @/metrics/<daemon_name>/<stat>.  This reduces the footprint of each
binary and makes the collected stats available in the cloud as well as
on the appliance.

Test Plan:
currently running on nils-wrt, with stats being read on
svc0.b10e.net

Differential Revision: https://ph0.b10e.net/D904