cl.http should die with a fatal error if it can't start HTTP or HTTPS services
Summary:
Production deployment of cl.httpd was failing because the HTTP service couldn't
start (because another process was using port 80), but the message didn't
include the error reason, and didn't cause the daemon to die.  The daemon
should exit if it can't start properly.

Test Plan:
Ran through several startup error conditions, saw that an error was properly
emitted, and that the daemon had a fatal error and exited.  There's
still a related issue when shutting down cleanly, which was documented.

Reviewers: duvall

Reviewed By: duvall

Differential Revision: https://ph0.b10e.net/D911