Reduce hotplug invocations
Summary:
When a new device or bridge is added or modified, the OpenWRT hotplug
framework will iterate over all of the network devices and redistribute
the interrupts for every rx and tx queue, attempting to spread the load
across all CPUs in the system.  When we tear down and rebuild the bridge
infrastucture, or restart hostapd, we will be iterating over the same
set of 35+ network devices dozens of times within a few seconds.  This
causes the system load to spike, which can lead to a failure spiral.

With this change, we will prevent the hotplug scripts from being invoked
during bridge rebuilds and hostapd restarts.  Instead, we will invoke
them a single time once the multi-step operation is completed.

Test Plan:
manual testing on nils-wrt and dogfood.  Tested across
reboots, restarts of ap.configd and ap.networkd, and killing hostapd.

Differential Revision: https://ph0.b10e.net/D917