Increase initial DHCP lease time
Summary:
When a client is initially assigned to a ring, we give it a 1 minute
DHCP lease.  On renewal it will get a full-length lease.  This is
intended to handle the race where the DHCP request arrives at
ap.serviced before ap.configd has assigned the client to the appropriate
ring.  When the initial short lease expires and the renewal request
arrives, ap.serviced will know which ring the client is assigned to.

This change increases that initial lease from 1 to 2 minutes.  Some DHCP
clients (including that on our OpenWRT boards) seem to have a minimum
renewal time of 1 minute, meaning that their renewal request always
arrives a second or two after that initial 1 minute lease expired.  As a
result, they never successfully renew their lease - they just get an
endless series of 1 minute leases.

Test Plan:
currently running on dogfood with nils-wrt as an OpenWRT
client.

Differential Revision: https://ph0.b10e.net/D916