Reduce scan load, particularly during network rebuilds
Summary:
Reduce the number of default password and vulnerabilty scan threads from
2 to 1.  This will reduce the CPU load on the system.

When networkd rebuilds the bridge infrastucture, pause scanning for a
minute.  This avoids competing for CPU with other daemons and the
hotplug framework.  It also avoids wasting time with scans that fail as
the connections they rely on are being torn down and rebuilt.

Test Plan: manual testing on nils-wrt.  Live system testing on dogfood.

Differential Revision: https://ph0.b10e.net/D919