Make config tree writes asynchronous
Summary:
Currently we marshal the config tree into json, write it to disk, and
sync the filesystem metadata before returning a completion to the
requesting daemon.  This increases command latency, increases system
load, and increases wear and tear on the flash.

With this change, the config tree is written to disk by an independent
go routine rather than being in the command completion path.  This
should dramatically reduce config latency.  In addition, changes will be
written to disk no more than once per second.  In an update-heavy
environment (which is typical at system startup), this will reduce
sytem load and flash wear.

Test Plan:
this has been running on nils-wrt and dogfood for the past
day or so.

Differential Revision: https://ph0.b10e.net/D918