improve hostapd stability by ignoring persistently bad clients
Summary:
If a client requires more than 6 RETRANSMIT operations within a 5 minute
window, we restart hostapd.  For some clients, this restart seems to
clear some bad state, allowing them to connect post-restart.  If a
client goes more than 5 minutes without a RETRANSMIT, we clear its
RETRANSMIT count, on the assumption that some corrective action may have
been taken.

Clients that are in some low-power state may go idle for long periods of
time, giving the impression that they have escaped the RETRANSMIT cycle.
When they wake up, the cycle begins again, and we will restart hostapd
again.  Because restarting hostapd is disruptive to all clients, this
should be avoided.

This change adds a new 'broken' flag to the retransmit state structure,
which will be set if the client ever exceeds its 6 ops in 5 minutes hard
limit.  That flag will remain set until a successful connection, even if
the count gets reset to zero.  The flag ensures that we will never
restart hostapd in an attempt to fix a client for which that has failed
in the past.

Test Plan:
currently running on dogfood with the persistently bad
7c:76:35:93:60:7a (HWLAB-INTELSTICK)

Differential Revision: https://ph0.b10e.net/D974