Use 'flattened tree' export to diagnose appliance/cloud config mismatches
Summary:
Add a Flatten() interface to the cfgtree module, which will convert a
property tree into a property-indexed map.  This flattened model is used
to compare the contents of the live config tree with the exported
version that we are about to upload to the cloud.  Any mismatch at this
point presages the persistent "hash mismatch" state in cl.configd.

Remove the comparison of live/file trees.  File trees are simply
persisted copies of the exported data, so the test was redundant.  In
addition, the switch to asynchronous persistence resulted in frequent
false positives.

Fix a caching bug in cfgtree, which was responsible for stale data being
uploaded to the cloud.  Full trees were being cached as "@", while
lookups were being performed on "@/".

Test Plan:
manual testing on nils-wrt, using extensive debug logging
which has been removed.  Deployed on dogfood and verified that the "hash
mismatch" errors on the cloud went away.

Differential Revision: https://ph0.b10e.net/D976