Allow cl.httpd to pull secrets from Vault
Summary:
The changes here allow `cl.httpd` to pull both key/value secrets (such as
external authentication credentials) and database credentials from a Vault
server, if the appropriate environment variables are set.  If the Vault server
is configured to rotate the database credentials, `cl.httpd` will reconnect with
the new credentials when that happens.

This also adds `application_name` to the database URI, which will show up in
places like the `pg_stat_activity` view.

Test Plan:
Run `cl.httpd` on a VM, with a single browser connection to it.  Unit tests for
the client code handling DB credential rotation.

Reviewers: dp

Reviewed By: dp

Differential Revision: https://ph0.b10e.net/D927