Decrease wifi scan frequency
Summary:
We periodically perform a full-spectrum scan on each wireless device,
trying to identify the least congested channels.  We use that
information to decide which channel we should use for our traffic.
Those scans are disruptive: as the wifi radio rapidly switches
frequencies, clients may detect the loss of signal and go through a
disconnect/reconnect cycle.  Apple devices seem to be particularly
sensitive to this disruption.

Eventually we should use the information about recent activity in
@/metrics/clients to avoid these scans when they are likely to be
disruptive.  For now, we will just significantly decrease the frequency
which which we perform the scans.

Test Plan:
running live on nils-plateu now.  Ping tests and watching the
logs for DISCONNECT/RECONNECT events confirms that client connections
are much more stable now.

Reviewers: dp

Differential Revision: https://ph0.b10e.net/D980