Fix cl-cert leaving behind lockfile
Summary:
`cl-cert` was calling `slog.Fatalw()` in a bunch of cases without
calling the method to remove the lockfile, resulting in future runs
being unable to get started.  This change makes sure to remove the
lockfile before exiting.

Test Plan:
Run resulting `cl-cert` on `svc1` and wait for errors; it seems to leave the
lockfile behind pretty often, so this shouldn't take too long.

Reviewers: dp, stephen

Reviewed By: stephen

Differential Revision: https://ph0.b10e.net/D977