fix 2 ap.tron bugs
Summary:
Fix a simultaneous map read/write by moving a Lock()/Unlock() pair out
of an 'if' clause.

Set the APGATEWAY environment variable before attempting to connect to
configd.  Without this, ap.tron on a satellite node will attempt to
connect to a non-existent config daemon at 127.0.0.1 instead of
connecting to the daemon on the gateway.  This bug prevents satellite
health info from being pushed into the config tree.

Test Plan: running on nils-wrt and nils-sat

Differential Revision: https://ph0.b10e.net/D982