config changes in preparation for vpn
Summary:
This adds CreateProps() and SetProps() to cfgapi, providing a simple
mechanism for a caller to update multiple properties at once.  This also
adds the notion of a 'null' property type, allowing the creation of a
property with no value.  Both of these are laying groundwork for the
forthcoming VPN code review(s).

Test Plan:
running on nils-wrt with wired and wireless clients, to
verify that both @/metrics consumers of the new APIs are being updated
correctly.

Differential Revision: https://ph0.b10e.net/D986