vpn user config management
Summary:
This adds the per-user vpn properties to the config tree.

This also adds ap-vpntool and cl-vpntool, which are used to generate and
manage per-user VPN keys.

Test Plan:
manual testing on nils-wrt and svc0.b10e.net.  Validated that
the full key lifecycle can be executed on the appliance, in the cloud,
and a mixture thereof.

Differential Revision: https://ph0.b10e.net/D988