cl.httpd should log to stackdriver
Summary:
This change brings `cl.httpd` in line with the other cloud services which
already log to Stackdriver, giving us a chance to use it in a dynamic, load
balanced environment without chasing down logs across multiple VMs.

Fixes T166

Test Plan:
Run in the load balanced, managed instance groups in the staging environment and
see that the logs exist and are useful.

Reviewers: dp

Reviewed By: dp

Maniphest Tasks: T166

Differential Revision: https://ph0.b10e.net/D991