Add vpn policies
Summary:
Allow the VPN functionality to be enabled and disabled at a site-wide
level.

Allow clients to route all traffic over the VPN - not just traffic to
the appliance's rings.

Make the vpn-related firewall rules dynamic.

The rings to which VPN clients are allowed access are specified in the
property tree rather than on the command line, ensuring that ap.networkd
has the information needed to construct the filtering firewall rules
correctly.

Added several convenience functions to cfgapi.

Test Plan: manual testing

Differential Revision: https://ph0.b10e.net/D1002