Replace persistent config pointer with opaque vpn handle
Summary:
The vpn module maintains some internal state to support config
callbacks and to reduce round trips to the config daemon.  This
optimization is incompatible with the way cl.httpd supports multiple
appliances.  To resolve the incompatibility, the state is now maintained
in an opaque structure which is returned as a side effect of vpn.Init()

Test Plan: manual testing on nils-wrt

Differential Revision: https://ph0.b10e.net/D1009